Refine 'unknown type' errors in function calls
